+++ tomato/src/index.js
@@ -4,24 +4,6 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyB9vD6srj0qUk79FWDUF-d6WwfxiIFM5N0",
-  authDomain: "tomato-648f6.firebaseapp.com",
-  projectId: "tomato-648f6",
-  storageBucket: "tomato-648f6.appspot.com",
-  messagingSenderId: "261646186918",
-  appId: "1:261646186918:web:b1210299fee5df3201f767"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
@@ -29,7 +11,4 @@ root.render(
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
